moved allcategories array in a separated file
