+++ frontend/src/components/ScrollableTabsButton.jsx
@@ -4,89 +4,12 @@ import Tab from "@mui/material/Tab";
 import Box from "@mui/material/Box";
 import { Avatar } from "@mui/material";
 import { useNavigate, useParams } from "react-router-dom";
+import allCategories from "../ressources/allCategories";
 
 export default function ScrollableTabsButton() {
-  const allCategories = [
-    {
-      id: 0,
-      cName: "American",
-      imgSrc:
-        "https://www.themealdb.com/images/media/meals/ryppsv1511815505.jpg/preview",
-    },
-    {
-      id: 1,
-      cName: "Italian",
-      imgSrc:
-        "https://www.themealdb.com/images/media/meals/1550441882.jpg/preview",
-    },
-    {
-      id: 2,
-      cName: "Japanese",
-      imgSrc:
-        "https://www.themealdb.com/images/media/meals/wpputp1511812960.jpg/preview",
-    },
-    // {
-    //   id: 3,
-    //   cName: "American",
-    //   imgSrc:
-    //     "https://www.themealdb.com/images/media/meals/ryppsv1511815505.jpg/preview",
-    // },
-    // {
-    //   id: 4,
-    //   cName: "Italian",
-    //   imgSrc:
-    //     "https://www.themealdb.com/images/media/meals/1550441882.jpg/preview",
-    // },
-    // {
-    //   id: 5,
-    //   cName: "Japanese",
-    //   imgSrc:
-    //     "https://www.themealdb.com/images/media/meals/wpputp1511812960.jpg/preview",
-    // },
-    // {
-    //   id: 6,
-    //   cName: "American",
-    //   imgSrc:
-    //     "https://www.themealdb.com/images/media/meals/ryppsv1511815505.jpg/preview",
-    // },
-    // {
-    //   id: 7,
-    //   cName: "Italian",
-    //   imgSrc:
-    //     "https://www.themealdb.com/images/media/meals/1550441882.jpg/preview",
-    // },
-    // {
-    //   id: 8,
-    //   cName: "Japanese",
-    //   imgSrc:
-    //     "https://www.themealdb.com/images/media/meals/wpputp1511812960.jpg/preview",
-    // },
-    // {
-    //   id: 9,
-    //   cName: "American",
-    //   imgSrc:
-    //     "https://www.themealdb.com/images/media/meals/ryppsv1511815505.jpg/preview",
-    // },
-    // {
-    //   id: 10,
-    //   cName: "Italian",
-    //   imgSrc:
-    //     "https://www.themealdb.com/images/media/meals/1550441882.jpg/preview",
-    // },
-    // {
-    //   id: 11,
-    //   cName: "Japanese",
-    //   imgSrc:
-    //     "https://www.themealdb.com/images/media/meals/wpputp1511812960.jpg/preview",
-    // },
-  ];
-
   const { category } = useParams();
 
-  const [value, setValue] = React.useState(
-    allCategories.filter((thisCategory) => thisCategory.cName === category)[0]
-      .id
-  );
+  const [value, setValue] = React.useState(null);
 
   const navigate = useNavigate();
 
@@ -94,7 +17,7 @@ export default function ScrollableTabsButton() {
     navigate(`/meals/${allCategories[newValue].cName}`);
     setValue(newValue);
   };
-
+  console.info(category);
   return (
     <Box sx={{ bgcolor: "background.paper" }}>
       <Tabs
